refactor(app): Select Zone Dropdown Component
