+++ app/components/select-zone.tsx
@@ -2,12 +2,10 @@
 
 import { MutableRefObject, useState, useRef } from 'react';
 import { BiddingZone } from '@/app/types';
-import { useMediaQuery } from '@/hooks/use-media-query';
-import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem, CommandList } from '@/components/ui/command';
-import { Drawer, DrawerContent, DrawerTrigger } from '@/components/ui/drawer';
-import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 import { HomeState, HomeController } from '@/app/home';
 import { LocationController, LocationSelectZone } from '@/components/select-zone-location';
+import { SelectZonesDropdown } from '@/components/select-zone-dropdown';
+import { BIDIGIT } from '@/app/constants';
 
 // State for the region select interface
 type RegionSelectState = {
@@ -22,27 +20,6 @@ export interface RegionSelectController {
   setBiddingZoneKeepLocation: (zone: BiddingZone) => void;
 }
 
-export const ZONES: BiddingZone[] = [
-  {
-    value: 'SE1',
-    label: 'North Sweden',
-  },
-  {
-    value: 'SE2',
-    label: 'North Central Sweden',
-  },
-  {
-    value: 'SE3',
-    label: 'South Central Sweden',
-  },
-  {
-    value: 'SE4',
-    label: 'South Sweden',
-  },
-];
-
-const bidigit: Intl.NumberFormat = new Intl.NumberFormat('en-US', { minimumIntegerDigits: 2 });
-
 export function RegionSelect({
   state,
   homeController,
@@ -52,9 +29,6 @@ export function RegionSelect({
   homeController: HomeController;
   controllerRef: MutableRefObject<RegionSelectController | null>;
 }) {
-  const [open, setOpen] = useState(false);
-  const isDesktop = useMediaQuery('(min-width: 768px)');
-
   const [controllerState, setControllerState] = useState<RegionSelectState>({
     zoneLoaded: false,
   });
@@ -80,73 +54,24 @@ export function RegionSelect({
   // Pass controller reference
   controllerRef.current = controller.current;
 
-  let dropdown;
-
   const dropdown_btn = (
     <div className="h-[3.5em] w-full">
       <div
         className={
           'text[#5A5A5A] flex h-full w-full cursor-pointer items-center justify-between whitespace-nowrap rounded-md border border-input bg-background p-[0.5em] text-[1em] text-sm font-medium leading-[1] shadow-sm transition-colors hover:bg-accent hover:text-accent-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50 dark:text-[#a3a3a3]' +
-          (controllerState.zoneLoaded ? ' !border-[#5164cd]' : '')
+          (controllerState.zoneLoaded ? ' !border-[hsl(var(--highlight))]' : '')
         }
       >
         <div className="font-[600] text-[#555]">{state.zone ? state.zone.value : 'ZONE'}</div>
         {state.zone ? state.zone.label : 'Select Zone'}
         <div className="font-[600] text-[#555]">
           {state.timeOfFetch
-            ? bidigit.format(state.timeOfFetch.getHours()) + ':' + bidigit.format(state.timeOfFetch.getMinutes())
+            ? BIDIGIT.format(state.timeOfFetch.getHours()) + ':' + BIDIGIT.format(state.timeOfFetch.getMinutes())
             : '--:--'}
         </div>
       </div>
     </div>
   );
-
-  const region_list = (
-    <Command>
-      <CommandInput placeholder="Filter zone..." />
-      <CommandList>
-        <CommandEmpty>No results found.</CommandEmpty>
-        <CommandGroup>
-          {ZONES.map((zone) => (
-            <CommandItem
-              key={zone.value}
-              value={zone.label}
-              onSelect={(value) => {
-                controller.current.setBiddingZone(ZONES.find((priority) => priority.label === value) || ZONES[0]);
-                setOpen(false);
-              }}
-            >
-              <div>
-                <span className="font-[600] text-[#555]">{zone.value}</span>
-                &nbsp;
-                {zone.label}
-              </div>
-            </CommandItem>
-          ))}
-        </CommandGroup>
-      </CommandList>
-    </Command>
-  );
-
-  if (isDesktop) {
-    dropdown = (
-      <Popover open={open} onOpenChange={setOpen}>
-        <PopoverTrigger asChild>{dropdown_btn}</PopoverTrigger>
-        <PopoverContent className="p-0" align="center">
-          {region_list}
-        </PopoverContent>
-      </Popover>
-    );
-  } else {
-    dropdown = (
-      <Drawer open={open} onOpenChange={setOpen}>
-        <DrawerTrigger asChild>{dropdown_btn}</DrawerTrigger>
-        <DrawerContent>
-          <div className="mt-4 border-t">{region_list}</div>
-        </DrawerContent>
-      </Drawer>
-    );
-  }
 
   return (
     <div className="flex w-[100%] max-w-[406px] gap-[5px] text-[14px]">
@@ -154,7 +79,7 @@ export function RegionSelect({
         controllerRef={locationSelectControllerRef}
         onSelectZone={controller.current.setBiddingZoneKeepLocation}
       />
-      {dropdown}
+      <SelectZonesDropdown onSelectZone={controller.current.setBiddingZone}>{dropdown_btn}</SelectZonesDropdown>
     </div>
   );
 }
